refactor: move duplicate code from both flows into flow_base.py
