+++ gladier_xpcs/flow_gpu.py
@@ -1,19 +1,15 @@
-import logging
-import time
-import random
-import globus_sdk
-from gladier import GladierBaseClient, generate_flow_definition, utils
+from gladier import generate_flow_definition
 # import gladier_xpcs.log  # Uncomment for debug logging
+from gladier_xpcs.flow_base import XPCSBaseClient
 
-
-log = logging.getLogger(__name__)
 
 @generate_flow_definition(modifiers={
    'publish_gather_metadata': {'payload': '$.GatherXpcsMetadata.details.result[0]'}
 })
-class XPCSGPUFlow(GladierBaseClient):
+class XPCSGPUFlow(XPCSBaseClient):
     globus_group = '368beb47-c9c5-11e9-b455-0efb3ba9a670'
     max_retries = 10
+    # See flow_base.py for assigning containers.
     containers = {}
 
     gladier_tools = [
@@ -25,39 +21,3 @@ class XPCSGPUFlow(GladierBaseClient):
         'gladier_xpcs.tools.gather_xpcs_metadata.GatherXPCSMetadata',
         'gladier_xpcs.tools.Publish',
     ]
-
-    def register_funcx_function(self, function):
-        """Register containers for any functions listed in self.containers"""
-        fxid_name = utils.name_generation.get_funcx_function_name(function)
-        if fxid_name in self.containers.keys():
-            container = self.containers[fxid_name]
-            log.info(f'Registering {fxid_name} with {container["location"]}')
-            fxid_name = utils.name_generation.get_funcx_function_name(function)
-            fxck_name = utils.name_generation.get_funcx_function_checksum_name(function)
-            cfg = self.get_cfg(private=True)
-            cid = self.funcx_client.register_container(**container)
-            fxid = self.funcx_client.register_function(function, function.__doc__, container_uuid=cid)
-            cfg[self.section][fxid_name] = fxid
-            cfg[self.section][fxck_name] = self.get_funcx_function_checksum(function)
-            cfg.save()
-        else:
-            super().register_funcx_function(function)
-
-    def retry_backoff(self, method, *args, **kwargs):
-        retries = 0
-        while retries < self.max_retries:
-            try:
-                return method(*args, **kwargs)
-            except globus_sdk.GlobusTimeoutError:
-                time.sleep(random.randint(1, 10))
-            except globus_sdk.GlobusAPIError as gapie:
-                if gapie.http_status == 400:
-                    raise
-                time.sleep(random.randint(1, 10))
-        raise
-
-    def run_flow(self, *args, **kwargs):
-        return self.retry_backoff(super().run_flow, *args, **kwargs)
-
-    def get_status(self, *args, **kwargs):
-        return self.retry_backoff(super().get_status, *args, **kwargs)
